Keep all logits in fp32
The vocabulary head wrote bf16 logits, which rounds near-ties together:
at a logit of 16 to 32 bf16 spacing is 0.125, and llama.cpp run with
its logits rounded to bf16 agrees on top-1 with its own f32 logits only
99.06% (27B) and 98.95% (35B) of the time. The loader now gives the head
an fp32 destination: its computation is unchanged (quantized weights,
bf16 input, tile, K splits, accumulation), and it stores its fp32 sums
instead of rounding them. Every place it runs writes fp32 logits: the
target's verify step at every lane count, the single-lane head of
prefill, restored prefixes and score-only requests, and the draft, which
shares the head. Both write the one Logits arena tensor, now sized for
fp32 (+16 MB for four lanes of the 248,320-token vocabulary); the
draft's rows are as many as the target's.

The sampling kernels read fp32 logits: the target's argmax and sharded
top-32 scans, and the draft selector's top-16 scan, which now streams
four tokens per 16-byte vector and keeps its candidates' scores in fp32.
The score logits of a request are its fp32 values, which the protocol
already carries as f32. draft-selector checks each scan, the
unconstrained greedy one included, against an fp32 CPU reference whose
order is decided below the bf16 spacing: a scan that reads the logits
rounded to bf16 fails it.

Against llama.cpp Metal's f32 logits (16,384 teacher-forced positions,
BF16 KV, M5 Pro), top-1 agreement rises from 98.91% to 99.44% on the 27B
and from 97.51% to 97.83% on the 35B, and nearly every flip it removes
lies within two bf16 ulps of a tie. Against the parent, a decode step
takes -1.7% to +0.4% on the 35B and 27B GGUF (M5 Pro, and the 35B on M3
Max), within ABBA noise, and +0.3% to +1.0% on the 35B MLX (M5 Pro, B1
to B4), where the doubled logits traffic shows; acceptance and outputs
are unchanged.

+++ runtime/ops/Sampling.cpp
@@ -51,7 +51,7 @@ DraftSelectorWorkspace Sampling::draftWorkspace(uint32_t positions) {
   // The partial values are followed by each position's 16 x 16 edge table.
   return {candidates * kDraftShards * sizeof(uint32_t),
           candidates * (kDraftShards + kDraftCandidates) * sizeof(float),
-          candidates * sizeof(uint32_t), candidates * sizeof(uint16_t),
+          candidates * sizeof(uint32_t), candidates * sizeof(float),
           candidates * sizeof(float)};
 }
 
@@ -98,7 +98,7 @@ void Sampling::addInitial(metal::CommandGraph &graph,
 
   metal::MetalBuffer logits = buffers.logits;
   if (rowOffset) {
-    const uint64_t rowBytes = uint64_t{vocabulary_} * sizeof(uint16_t);
+    const uint64_t rowBytes = uint64_t{vocabulary_} * sizeof(float);
     logits = backend_.view(logits, uint64_t{rowOffset} * rowBytes, rowBytes);
   }
   graph.add("decode_sample_argmax_sharded",
